Move CompactLength into it's own module
