+++ src/style/mod.rs
@@ -1,6 +1,7 @@
 //! A typed representation of [CSS style properties](https://css-tricks.com/snippets/css/a-guide-to-flexbox/) in Rust. Used as input to layout computation.
 mod alignment;
 mod available_space;
+mod compact_length;
 mod dimension;
 
 #[cfg(feature = "block_layout")]
@@ -12,7 +13,8 @@ mod grid;
 
 pub use self::alignment::{AlignContent, AlignItems, AlignSelf, JustifyContent, JustifyItems, JustifySelf};
 pub use self::available_space::AvailableSpace;
-pub use self::dimension::{CompactLength, Dimension, LengthPercentage, LengthPercentageAuto};
+pub use self::compact_length::CompactLength;
+pub use self::dimension::{Dimension, LengthPercentage, LengthPercentageAuto};
 
 #[cfg(feature = "block_layout")]
 pub use self::block::{BlockContainerStyle, BlockItemStyle, TextAlign};
